refactor: move _handle_error to base class
to review: deep research, firecrawl, metasleuth, twitter profile,
zkignite
