+++ mesh/agents/allora_price_prediction_agent.py
@@ -119,12 +119,6 @@ class AlloraPricePredictionAgent(MeshAgent):
             temperature=temperature,
         )
 
-    def _handle_error(self, maybe_error: dict) -> dict:
-        """Check for and return any errors in the response"""
-        if "error" in maybe_error:
-            return {"error": maybe_error["error"]}
-        return {}
-
     # ------------------------------------------------------------------------
     #                      ALLORA API-SPECIFIC METHODS
     # ------------------------------------------------------------------------
