+++ mesh/agents/carv_onchain_data_agent.py
@@ -165,15 +165,6 @@ class CarvOnchainDataAgent(MeshAgent):
             temperature=temperature,
         )
 
-    def _handle_error(self, maybe_error: dict) -> dict:
-        """
-        Small helper to return the error if present in
-        a dictionary with the 'error' key.
-        """
-        if "error" in maybe_error:
-            return {"error": maybe_error["error"]}
-        return {}
-
     # ------------------------------------------------------------------------
     #                      CARV API-SPECIFIC METHODS
     # ------------------------------------------------------------------------
