+++ mesh/agents/coingecko_token_info_agent.py
@@ -872,15 +872,6 @@ class CoinGeckoTokenInfoAgent(MeshAgent):
             temperature=temperature,
         )
 
-    def _handle_error(self, maybe_error: dict) -> dict:
-        """
-        Small helper to return the error if present in
-        a dictionary with the 'error' key.
-        """
-        if "error" in maybe_error:
-            return {"error": maybe_error["error"]}
-        return {}
-
     @monitor_execution()
     @with_retry(max_retries=3)
     async def handle_message(self, params: Dict[str, Any]) -> Dict[str, Any]:
