+++ mesh/agents/exa_search_agent.py
@@ -146,15 +146,6 @@ class ExaSearchAgent(MeshAgent):
             temperature=temperature,
         )
 
-    def _handle_error(self, maybe_error: dict) -> dict:
-        """
-        Small helper to return the error if present in
-        a dictionary with the 'error' key.
-        """
-        if "error" in maybe_error:
-            return {"error": maybe_error["error"]}
-        return {}
-
     # ------------------------------------------------------------------------
     #                      EXA API-SPECIFIC METHODS
     # ------------------------------------------------------------------------
