+++ mesh/agents/funding_rate_agent.py
@@ -189,15 +189,6 @@ class FundingRateAgent(MeshAgent):
             temperature=temperature,
         )
 
-    def _handle_error(self, maybe_error: dict) -> dict:
-        """
-        Small helper to return the error if present in
-        a dictionary with the 'error' key.
-        """
-        if "error" in maybe_error:
-            return {"error": maybe_error["error"]}
-        return {}
-
     # ------------------------------------------------------------------------
     #                      COINSIDER API-SPECIFIC METHODS
     # ------------------------------------------------------------------------
