+++ mesh/agents/masa_twitter_search_agent.py
@@ -135,15 +135,6 @@ class MasaTwitterSearchAgent(MeshAgent):
             temperature=temperature,
         )
 
-    def _handle_error(self, maybe_error: dict) -> dict:
-        """
-        Small helper to return the error if present in
-        a dictionary with the 'error' key.
-        """
-        if "error" in maybe_error:
-            return {"error": maybe_error["error"]}
-        return {}
-
     # ------------------------------------------------------------------------
     #                      MASA API-SPECIFIC METHODS
     # ------------------------------------------------------------------------
