+++ mesh/agents/moni_twitter_insight_agent.py
@@ -197,15 +197,6 @@ class MoniTwitterInsightAgent(MeshAgent):
             temperature=temperature,
         )
 
-    def _handle_error(self, maybe_error: dict) -> dict:
-        """
-        Small helper to return the error if present in
-        a dictionary with the 'error' key.
-        """
-        if "error" in maybe_error:
-            return {"error": maybe_error["error"]}
-        return {}
-
     def _clean_username(self, username: str) -> str:
         """
         Remove @ symbol if present in the username
