+++ mesh/agents/sol_wallet_agent.py
@@ -196,15 +196,6 @@ class SolWalletAgent(MeshAgent):
             temperature=temperature,
         )
 
-    def _handle_error(self, maybe_error: dict) -> dict:
-        """
-        Small helper to return the error if present in
-        a dictionary with the 'error' key.
-        """
-        if "error" in maybe_error:
-            return {"error": maybe_error["error"]}
-        return {}
-
     # ------------------------------------------------------------------------
     #                      HELIUS API-SPECIFIC METHODS
     # ------------------------------------------------------------------------
